requirements.txt
backcall==0.1.0
certifi==2019.11.28
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
kiwisolver==1.1.0
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nltk==3.4.5
numpy==1.18.1
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.4
Pygments==2.6.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
textblob==0.15.3
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
wincertstore==0.2
